Language.Sunroof.Classes

Description

Provides the central type classes used by Sunroof.

Documentation

class Sunroof a where

Central type class of Sunroof. Every type that can be translated into Javascript with Sunroof has to implement this type class.

Minimal complete definition

box, unbox

Methods

box :: Expr -> a

Create a Sunroof value from a plain Javascript expression.

unbox :: a -> Expr

Reveal the plain Javascript expression that represents this Sunroof value.

typeOf :: Proxy a -> Type

Returns the type of Javascript expression this Sunroof value represents. The default implementation returns Base as type.

class SunroofValue a where

All Haskell values that have a Sunroof representation implement this class.

Associated Types

type ValueOf a :: *

The Sunroot type that is equivalent to the implementing Haskell type.

Methods

js :: a -> ValueOf a

Convert the Haskell value to its Sunroof equivalent.

class SunroofArgument args where

Everything that can be used as argument to a function is Javascript/Sunroof.

Methods

jsArgs :: args -> [Expr]

Turn the argument into a list of expressions.

jsValue :: UniqM m => m args

Create a list of fresh variables for the arguments.

typesOf :: Proxy args -> [Type]

Get the type of the argument values.

class Monad m => UniqM m where

Implemented if a monad supports unique number generation.

Methods

uniqM :: m Uniq

Generate a unique number.

type Uniq = Int

Used for unique number generation.

mkVar :: Sunroof a => Uniq -> a

Creates a Javascript variable of any Sunroof type.

jsVar :: (Sunroof a, UniqM m) => m a

Create a unique Javascript variable of any Sunroof type.