Language.Sunroof.JS.Browser

Description

A reflection of the standard browser Javascript API.

Top level functions

alert :: JSString -> JS t ()

Display the given text in a message box.

See http://www.w3schools.com/js/js_popup.asp.

confirm :: JSString -> JS t JSBool

Ask the user to confirm the given massege.

See http://www.w3schools.com/js/js_popup.asp.

prompt :: JSString -> JSString -> JS t JSObject

Ask the user to give some input. May return null or a string. Don't forget to check against nullJS before casting to string.

See http://www.w3schools.com/js/js_popup.asp.

decodeURI :: JSString -> JS t JSString

Decode the URI encoded in the given string.

encodeURI :: JSString -> JS t JSString

Encode the given string in URI encoding.

decodeURIComponent :: JSString -> JS t JSString

Decode the URI encoded string. For use with encodeURIComponent.

encodeURIComponent :: JSString -> JS t JSString

Encode the string with URI encoding. This encodes a few more characters to make the string safe for direct server communication (AJAX).

eval :: Sunroof a => JSString -> JS t a

Evaluate the given JavaScript string if possible. Returns the result of evaluation. TODO: think about this a bit.

isFinite :: JSNumber -> JS t JSBool

Check if a given number is within the valid JavaScript number range.

isNaN :: JSNumber -> JS t JSBool

Check if a given number is NaN or not.

parseFloat :: JSString -> JS t JSNumber

Parse the given string to a number.

parseInt :: JSString -> JS t JSNumber

Parse the given string to a number.

Window API

window :: JSObject

The window object.

setInterval :: (() -> JSB ()) -> JSNumber -> JSObject -> JS t JSNumber

Calls a function at specified intervals in milliseconds. It will continue calling the function until clearInterval is called, or the window is closed. The returned number is needed for clearInterval. This is supposed to be called on the window object. See: http://www.w3schools.com/jsref/met_win_setinterval.asp

clearInterval :: JSNumber -> JSObject -> JS t ()

Clears a timer set with the setInterval method. This is supposed to be called on the window object. See: http://www.w3schools.com/jsref/met_win_clearinterval.asp

setTimeout :: (() -> JSB ()) -> JSNumber -> JSObject -> JS t JSNumber

Execute the given continutation after the given amount of milliseconds. Returns a handler for the set timer. This is supposed to be called on the window object. See: http://www.w3schools.com/jsref/met_win_settimeout.asp

clearTimeout :: JSNumber -> JSObject -> JS t ()

Removes the timer associated with the given handler. This is supposed to be called on the window object. See: http://www.w3schools.com/jsref/met_win_cleartimeout.asp

Screen API

screen :: JSObject

The screen object.

Document API

document :: JSObject

The document object.

getElementById

Arguments

:: JSString
The id.

-> JSObject

-> JS t JSObject

Get the DOM object of the element with the given id. For use with document.

getElementsByName

Arguments

:: JSString
The name.

-> JSObject

-> JS t JSObject

Get the DOM objects of the elements with the given name. For use with document.

getElementsByTagName

Arguments

:: JSString
The tag name.

-> JSObject

-> JS t JSObject

Get the DOM objects of the elements with the given tag. For use with document.

createAttribute

Arguments

:: JSString
The name of the new attribute.

-> JSObject

-> JS t JSObject

Create a attribute DOM node with the given name. For use with document.

createElement

Arguments

:: JSString
The tag name of the new element.

-> JSObject

-> JS t JSObject

Create a element DOM node with the given tag name. For use with document.

createTextNode

Arguments

:: JSString
The text of the new text node.

-> JSObject

-> JS t JSObject

Create a text DOM node with the given string as text. For use with document.

open :: JSObject -> JS t ()

Opens the document for writing. For use with document.

close :: JSObject -> JS t ()

Closes the document after writing. For use with document.

write :: JSString -> JSObject -> JS t ()

Writes something into the document. For use with document.

writeln :: JSString -> JSObject -> JS t ()

Write something into the document and appends a new line. For use with document.

setCookie :: JSString -> JSObject -> JS t ()

Sets the value of the cookie. For use with document.

cookie :: JSSelector JSString

Returns the value of the cookie. For use with document.

referrer :: JSSelector JSString

Returns the referrer of the document. For use with document.

setTitle :: JSString -> JSObject -> JS t ()

Sets the title of the document. For use with document.

title :: JSSelector JSString

Returns the title of the document. For use with document.

url :: JSSelector JSString

Returns the complete URL of the document. For use with document.

Image DOM

src :: JSSelector JSString

Returns the src of a DOM image object.

The JavaScript Console

data JSConsole

The type of the debugging console object. See: https://developers.google.com/chrome-developer-tools/docs/console-api, https://developer.mozilla.org/en-US/docs/DOM/console, http://msdn.microsoft.com/en-us/library/windows/apps/hh696634.aspx;

console :: JSConsole

The console object.

log :: SunroofArgument a => a -> JSConsole -> JS t ()

Log the given message.

debug :: SunroofArgument a => a -> JSConsole -> JS t ()

Send a debug level message to the console.

info :: SunroofArgument a => a -> JSConsole -> JS t ()

Send a info message to the console.

warn :: SunroofArgument a => a -> JSConsole -> JS t ()

Send a warning message to the console.

error :: SunroofArgument a => a -> JSConsole -> JS t ()

Send an error message to the console.