Language.Sunroof.JS.Chan

Description

JSChan provides the same functionality and concurrency abstraction in Javascript computations as Chan in Haskell.

Documentation

data JSChan a

JSChan abstraction. The type parameter gives the type of values held in the channel.

newChan :: SunroofArgument a => JS t (JSChan a)

Create a new empty JSChan.

writeChan :: forall t a. (SunroofThread t, SunroofArgument a) => a -> JSChan a -> JS t ()

Put a value into the channel. This will never block.

readChan :: forall a. (Sunroof a, SunroofArgument a) => JSChan a -> JS B a

Take a value out of the channel. If there is no value inside, this will block until one is available.