Language.Sunroof.JS.JQuery

Description

This module provides parts of the JQuery API (http://api.jquery.com/).

General JQuery API

dollar :: JSFunction JSString JSObject

The dollar function. See http://api.jquery.com/jQuery/.

jQuery :: JSString -> JS t JSObject

Calls the JQuery dollar function. See http://api.jquery.com/jQuery/.

jq :: JSString -> JS t JSObject

Short-hand for jQuery.

DOM

append :: JSObject -> JSObject -> JS t ()

See http://api.jquery.com/append/.

html :: JSObject -> JS t JSObject

See .html() at http://api.jquery.com/html/.

setHtml :: JSString -> JSObject -> JS t JSObject

See .html(htmlString) at http://api.jquery.com/html/.

text :: JSObject -> JS t JSObject

See .text() at http://api.jquery.com/text/.

setText :: JSString -> JSObject -> JS t JSObject

See .text(textString) at http://api.jquery.com/text/.

CSS

css :: JSString -> JSObject -> JS t JSString

See .css(propertyName) at http://api.jquery.com/css/.

setCss :: JSString -> JSString -> JSObject -> JS t JSString

See .css(propertyName, value) at http://api.jquery.com/css/.

addClass :: JSString -> JSObject -> JS t ()

See http://api.jquery.com/addClass/.

removeClass :: JSString -> JSObject -> JS t ()

See http://api.jquery.com/removeClass/.

Attributes

attribute :: JSString -> JSObject -> JS t JSString

See .attr(attributeName) at http://api.jquery.com/attr/. This binding does not have the original Javascript name, because of the attr function.

attr' :: JSString -> JSObject -> JS t JSString

See .attr(attributeName) at http://api.jquery.com/attr/. This binding does not have the original Javascript name, because of the attr function.

setAttr :: JSString -> JSString -> JSObject -> JS t JSString

See .attr(attributeName, value) at http://api.jquery.com/attr/.

removeAttr :: JSString -> JSObject -> JS t JSObject

See: http://api.jquery.com/removeAttr/

Event Handling

on :: SunroofArgument a => JSString -> JSString -> (a -> JS B ()) -> JSObject -> JS t ()

See http://api.jquery.com/on/.

Manipulation

innerWidth :: JSObject -> JS t JSNumber

See http://api.jquery.com/innerHeight/.

innerHeight :: JSObject -> JS t JSNumber

See http://api.jquery.com/innerWidth/.

outerWidth :: JSObject -> JS t JSNumber

See http://api.jquery.com/outerWidth/.

outerWidth' :: JSBool -> JSObject -> JS t JSNumber

See http://api.jquery.com/outerWidth/.

outerHeight :: JSObject -> JS t JSNumber

See http://api.jquery.com/outerHeight/.

outerHeight' :: JSBool -> JSObject -> JS t JSNumber

See http://api.jquery.com/outerHeight/.

clone :: JSObject -> JS t JSObject

See .clone() at http://api.jquery.com/clone/.

clone' :: JSBool -> JSBool -> JSObject -> JS t JSObject

See .clone(withDataAndEvents, deepWithDataAndEvents) at http://api.jquery.com/clone/.