Language.Sunroof.JS.MVar

Description

JSMVar provides the same functionality and concurrency abstraction in Javascript computations as MVar in Haskell.

Documentation

data JSMVar a

JSMVar abstraction. The type parameter gives the type of values held in a JSMVar.

newMVar :: forall a t. SunroofArgument a => a -> JS t (JSMVar a)

Create a new JSMVar with the given value inside. See newEmptyMVar.

newEmptyMVar :: SunroofArgument a => JS t (JSMVar a)

Create a new empty JSMVar. See newMVar.

putMVar :: forall a. SunroofArgument a => a -> JSMVar a -> JS B ()

Put the value into the JSMVar. If there already is a value inside, this will block until it is taken out.

takeMVar :: forall a. (Sunroof a, SunroofArgument a) => JSMVar a -> JS B a

Take the value out of the JSMVar. If there is no value inside, this will block until one is available.