Language.Sunroof.JS.Map

Description

JSMap provides an abstract and more type-safe access to maps in JavaScript. It is a wrapper around the dictionary each object in JavaScript is.

Documentation

data JSMap k a

JSMap abstraction. The first type parameter gives the type of keys used by the name and the second gives the type of values.

newMap :: JS t (JSMap k a)

Create a new empty JSMap.

insert :: (SunroofKey k, Sunroof a) => k -> a -> JSMap k a -> JS t ()

insert k x inserts an element x associated with the given key k into a map.

lookup' :: (SunroofKey k, Sunroof a) => k -> JSMap k a -> JS t a

lookup k selects the value associated with the key k.

size :: JSMap k a -> JS t JSNumber